backend/douyin_gpt/requirements.txt
streamlit>=1.30.0
apify-client>=1.6.0
openai-whisper>=20231117
python-docx>=1.1.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
douyin-tiktok-scraper>=1.6.0
